Separate Interpeter implementations into separate source file.
