+++ Makefile
@@ -200,6 +200,7 @@ INTERP_SRCS_BASE = \
 	ByteWriter.cpp \
 	ByteWriteStream.cpp \
 	DecompressSelector.cpp \
+	Interpreter.cpp \
 	IntFormats.cpp \
 	IntReader.cpp \
 	IntStream.cpp \
